test(checkbox): recategorize and clean up browser tests
